philo_one

+++ Philosophers/philo_one/srcs/ft_utils.c
@@ -5,7 +5,7 @@ suseconds_t	ft_get_time(void)
 	struct timeval time;
 
 	gettimeofday(&time, NULL);
-	return ((long)time.tv_sec * 1000 + time.tv_usec / 1000);
+	return (time.tv_sec * 1000 + time.tv_usec / 1000);
 }
 
 suseconds_t	ft_time_diff(t_m *mn)
